refactor: extract mockRouteDeps for reuse (validate-proxy)
Move the mock RouteDeps builder from scripts/validate-routes.ts into
ingest/mock-route-deps.ts so the next change (check-proxy → validate-proxy
via createRoutes) can reuse it. No behavior change — validate-routes now
imports mockRouteDeps() instead of defining it inline.

This unblocks the follow-up to make validate-proxy derive expected
DEV_PROXY_PATHS from the live route table (same approach as
validate-routes), as noted in notes/validate-proxy.md.

Verified: tsc pass, validate-routes OK (40 routes → 54 events), 859 tests.
